fix(search): Use server component
