+++ apps/web/src/common/utils/next13.ts
@@ -1,5 +1,8 @@
-export interface Next13ServerPageProps<Param extends string> {
+export interface Next13ServerPageProps<Param extends string = "", SearchParams extends string = ""> {
   params: {
     [T in Param]: string;
+  };
+  searchParams: {
+    [T in SearchParams]: string;
   };
 }
